+++ src/app/page.tsx
@@ -1,40 +1,5 @@
-"use client";
 import Image from 'next/image';
-import { Button, Navbar, NavbarBrand, NavbarCollapse, NavbarLink, NavbarToggle } from "flowbite-react";
-
-
-
-// import { Button, Navbar } from "flowbite-react";
-// className="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto md:w-auto p-4"
-
-type childSign = {
-  name: string,
-}
-export function ComponentNav(propC: childSign) {
-  return (
-    <Navbar fluid className='bg-pallete-1 items-center justify-between mx-auto md:w-auto p-3'>
-      <Navbar.Brand href="http://localhost:3000/csc-324-sga-assisting-tool">
-        <span className="self-center whitespace-nowrap text-xl font-semibold text-black">GBudget</span>
-      </Navbar.Brand>
-      <div className="flex md:order-2">
-        <a href="/csc-324-sga-assisting-tool/help"  className="flex items-center space-x-3 rtl:space-x-reverse"> 
-          <Button type="button" className="text-left text-black bg-pallete-1 font-medium text-sm order-1">Help</Button>
-        </a>
-        <Button className='bg-pallete-5 font-medium text-sm w-28 h-10'>{propC.name}</Button>
-        <Navbar.Toggle />
-      </div>
-      <Navbar.Collapse className="justify-between hidden w-full md:flex md:w-auto md:order-1">
-        <Navbar.Link href="#" active className='text-black text-base font-medium'>
-          Home
-        </Navbar.Link>
-        <Navbar.Link href="#" className='text-black text-base font-medium'>About</Navbar.Link>
-        <Navbar.Link href="#" className='text-black text-base font-medium'>Services</Navbar.Link>
-        <Navbar.Link href="#" className='text-black text-base font-medium'>Pricing</Navbar.Link>
-        <Navbar.Link href="#" className='text-black text-base font-medium'>Contact</Navbar.Link>
-      </Navbar.Collapse>
-    </Navbar>
-  );
-}
+import { ComponentNav } from './navbarComp';
 
 export default function Home() {
   return (
@@ -42,7 +7,7 @@ export default function Home() {
       {/* <h1 className="text-3xl font-bold underline">
         Hello world!
       </h1> */}
-      <ComponentNav name ="Log Out"/>
+      {/* <ComponentNav name ="Log Out"/> */}
       
 
     </main>
@@ -160,43 +125,3 @@ export default function Home() {
 //     </main>
 //   );
 // }
-
-
-
-
-// <nav className="bg-pallete-1 border-gray-200 dark:bg-gray-900">
-//         <div className="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto md:w-auto p-4">
-//           <a href="http://localhost:3000/csc-324-sga-assisting-tool" className="flex items-center space-x-3 rtl:space-x-reverse">
-//               <img src="https://flowbite.com/docs/images/logo.svg" className="h-8" alt="Flowbite Logo" />
-//               <span className="text-black self-center text-2xl font-semibold whitespace-nowrap dark:text-black">GBudget</span>
-//           </a>
-//           <div className="flex md:order-2 space-x-3 md:space-x-0 rtl:space-x-reverse">
-//             <a href="/help"  className="flex items-center space-x-3 rtl:space-x-reverse"> 
-//               <button type="button" className="text-left text-black bg-pallete-1 hover:bg-blue-800  font-medium text-sm order-1">Help</button>
-//             </a>
-//             <button type="button" className="text-white bg-pallete-5 order-2 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">Log Out</button>
-//             {/* <button data-collapse-toggle="navbar-cta" type="button" className="inline-flex items-center p-2 w-10 h-10 justify-center text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600" aria-controls="navbar-cta" aria-expanded="false">
-//               <span className="sr-only">Open main menu</span>
-//               <svg className="w-5 h-5" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 17 14">
-//                   <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M1 1h15M1 7h15M1 13h15"/>
-//               </svg>
-//             </button> */}
-//           </div>
-//           <div className="justify-between hidden w-full md:flex md:w-auto md:order-1" id="navbar-cta">
-//             <ul className="flex flex-col font-medium p-4 md:p-0 mt-4 border border-gray-100 rounded-lg bg-gray-50 md:space-x-8 rtl:space-x-reverse md:flex-row md:mt-0 md:border-0 md:bg-pallete-1 ">
-//               <li>
-//                 <a href="#" className="block py-2 px-3 md:p-0 text-black bg-blue-700 rounded md:bg-transparent md:text-blue-700 md:dark:text-blue-500" aria-current="page">Home</a>
-//               </li>
-//               <li>
-//                 <a href="#" className="block py-2 px-3 md:p-0 text-gray-900 rounded hover:bg-gray-100 md:hover:bg-transparent md:hover:text-blue-700 md:dark:hover:text-blue-500 dark:text-white dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent dark:border-gray-700">About</a>
-//               </li>
-//               <li>
-//                 <a href="#" className="block py-2 px-3 md:p-0 text-gray-900 rounded hover:bg-gray-100 md:hover:bg-transparent md:hover:text-blue-700 md:dark:hover:text-blue-500 dark:text-white dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent dark:border-gray-700">Services</a>
-//               </li>
-//               <li>
-//                 <a href="#" className="block py-2 px-3 md:p-0 text-gray-900 rounded hover:bg-gray-100 md:hover:bg-transparent md:hover:text-blue-700 md:dark:hover:text-blue-500 dark:text-white dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent dark:border-gray-700">Contact</a>
-//               </li>
-//             </ul>
-//           </div>
-//         </div>
-//       </nav>
